tests/wpt/web-platform-tests/webmessaging/with-options/host-specific-origin.tentative.html

<!doctype html>
<title>resolving url with stuff in host-specific</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  postMessage('', [], {targetOrigin: location.protocol + '//' + location.host + '//'});
  onmessage = this.step_func(function(e) {
    assert_equals(e.origin, location.protocol + '//' + location.host);
    this.done();
  });
});
</script>
